% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAT.R
\docType{data}
\name{biblio}
\alias{biblio}
\title{biblio file for testing.}
\format{
A data.frame with bibliographical data.
}
\usage{
data(biblio)
}
\description{
A dataset from Web of Science, exported as tab delimited text, full record.
}
\keyword{datasets}
