% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAT.R
\docType{data}
\name{map}
\alias{map}
\title{Matrix matching country names and coordinates.}
\format{
A matrix with countries and corresponding coordinates.
}
\usage{
data(map)
}
\description{
A dataset that links author countries with the map using the coordinates.
}
\keyword{datasets}
