% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rat_bar.R
\name{rat_bar}
\alias{rat_bar}
\title{A simple bar chart with one variable}
\usage{
rat_bar(df, col)
}
\arguments{
\item{col}{}
}
\value{

}
\description{
A simple bar chart with one variable
}
\examples{
rat_bar(jj_data, age_group)

jj_data \%>\%
   rat_bar(age_group)
}
