% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{RAT_feedback_with_graph}
\alias{RAT_feedback_with_graph}
\title{RAT feedback (with graph)}
\usage{
RAT_feedback_with_graph(dict = RAT::RAT_dict)
}
\arguments{
\item{dict}{The psychTestR dictionary used for internationalisation.}
}
\description{
Here the participant is given textual and graphical feedback at the end of the test.
}
\examples{
\dontrun{
RAT_demo(feedback = RAT_feedback_with_score())}
}
