% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{RAT_feedback_with_score}
\alias{RAT_feedback_with_score}
\title{RAT feedback (no score)}
\usage{
RAT_feedback_with_score(dict = RAT::RAT_dict)
}
\arguments{
\item{dict}{The psychTestR dictionary used for internationalisation.}
}
\description{
Here the participant is given no feedback at the end of the test.
}
\examples{
\dontrun{
RAT_demo(feedback = RAT_feedback_no_score())}
}
