% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rat_stacked_bar.R
\name{rat_stacked_bar}
\alias{rat_stacked_bar}
\title{A stacked bar chart.}
\usage{
rat_stacked_bar(df, col, group_col)
}
\arguments{
\item{df}{a dataframe}

\item{col}{x-axis variable (numeric)}

\item{group_col}{color fill variable (categorical)}
}
\value{

}
\description{
A stacked bar chart.
}
\examples{
rat_stacked_bar(jj_data, report_year, race)

jj_data \%>\%
   rat_stacked_bar(report_year, race)
}
