% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAT.R
\name{wos}
\alias{wos}
\title{Web of Science.}
\usage{
wos(
  id,
  user = NULL,
  pass = NULL,
  coll = c("SCI", "SSCI", "AHCI", "ISTP", "ISSHP", "BSCI", "BHCI", "ESCI")
)
}
\arguments{
\item{id}{ResearcherID or name of the researcher as 'Surname, First name'. Beware that in case of multiple authors with the same name all publications will be returned. In alternative, use the search notation of Web of Science for regular searches beyond an author's name.}

\item{user}{username for Web of Science. Not needed if you have access through your institution.}

\item{pass}{password or Web of Science. Not needed if you have access through your institution.}

\item{coll}{Web of Science collections to include.}
}
\value{
A list with publication data.
}
\description{
Downloads data from Web of Science.
}
\details{
The data downloaded can be used in h.index and i.index functions to avoid constant downloads.
}
\examples{
wos("C-2482-2012")
}
