% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAT.R
\name{i.index}
\alias{i.index}
\title{I-index.}
\usage{
i.index(id, fulldata = FALSE)
}
\arguments{
\item{id}{A list obtained with function 'wos'.}

\item{fulldata}{if TRUE returns publication and citation counts.}
}
\value{
The i-index value. If fulldata = TRUE a list with full data.
}
\description{
Calculates the i-index based on Web of Science data.
}
\details{
The i-index is a measure of scientific collaborations across countries. Calculated as the i number of co-author countries in more than i papers (Cardoso et al. subm.).
}
\examples{
id = wos("C-2482-2012")
i.index(id, fulldata = TRUE)
}
\references{
Cardoso, P., Fukushima, C.S. & Mammola, S. (subm.) Quantifying the international collaboration attitude of scholars.
}
